Rename package to etched_peakrdl_cheader.
