+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "peakrdl-cheader"
+name = "etched-peakrdl-cheader"
 dynamic = ["version"]
 requires-python = ">=3.6"
 dependencies = [
@@ -48,13 +48,13 @@ Changelog = "https://github.com/SystemRDL/PeakRDL-cheader/releases"
 Documentation = "https://peakrdl-cheader.readthedocs.io/"
 
 [tool.setuptools.dynamic]
-version = {attr = "peakrdl_cheader.__about__.__version__"}
+version = {attr = "etched_peakrdl_cheader.__about__.__version__"}
 
 [project.entry-points."peakrdl.exporters"]
-c-header = "peakrdl_cheader.__peakrdl__:Exporter"
+c-header = "etched_peakrdl_cheader.__peakrdl__:Exporter"
 
 [tool.setuptools]
 include-package-data = true
 
 [tool.setuptools.package-data]
-"peakrdl_cheader" = ["templates/*"]
+"etched_peakrdl_cheader" = ["templates/*"]
